Data save for current keyboard (#377)

+++ src/renderer/App.js
@@ -161,7 +161,7 @@ class App extends React.Component {
     if (isDemo) {
       this.setState({
         connected: true,
-        device: null
+        device: port
       });
       await navigate("/demoeditor");
       return;
@@ -195,13 +195,14 @@ class App extends React.Component {
   };
 
   onKeyboardDisconnect = async () => {
+    const { isDemo } = this.state;
     focus.close();
     this.setState({
       connected: false,
       device: null,
       pages: {}
     });
-    localStorage.clear();
+    if (!isDemo) localStorage.clear();
     await navigate("/keyboard-select");
   };
 
@@ -232,7 +233,7 @@ class App extends React.Component {
     let focus = new Focus();
     let device =
       (focus.device && focus.device.info) ||
-      (this.state.device && this.state.device.info);
+      (this.state.device && this.state.device.device.info);
 
     return (
       <MuiThemeProvider theme={this.state.darkMode ? darkTheme : lightTheme}>
@@ -277,6 +278,7 @@ class App extends React.Component {
                   startContext={this.startContext}
                   cancelContext={this.cancelContext}
                   inContext={this.state.contextBar}
+                  device={this.state.device}
                   titleElement={() => document.querySelector("#page-title")}
                   appBarElement={() => document.querySelector("#appbar")}
                 />
